Code review fixes: return errors from trigger API, fix edge cases
- All ds_trigger_* functions now return int (0/-1) instead of void
- Add stdbool.h include, fix uint8_t slope subtraction overflow
- Rollback BT seq_tag on write failure
- Remove dead code (read_uevent_hex)
- Enable -Wall -Wextra -Wpedantic, fix truncation warnings
- Add CMAKE_EXPORT_COMPILE_COMMANDS, sanitizer option
- Add install.sh one-liner script
- Document thread safety and ds_send() requirement in header

+++ src/triggers.c
@@ -92,7 +92,7 @@ bool ds_effect_slope_feedback(uint8_t *dst, uint8_t start_pos, uint8_t end_pos,
 		return false;
 
 	uint8_t strengths[10] = {0};
-	float slope = (float)(end_str - start_str) / (float)(end_pos - start_pos);
+	float slope = (float)((int)end_str - (int)start_str) / (float)((int)end_pos - (int)start_pos);
 
 	for (int i = start_pos; i < 10; i++) {
 		if (i <= end_pos)
